Experimental passing proxies as in and inout args
If CXXSWIZZLE_VECTOR_INOUT_WRAPPER_ENABLED is defined, proxies will also
be convertible to vector_inout_wrapper, which on destruction copies its
value back into the proxy. Potentially slow (uses std::function
internally), but enabled use of inout and in with proxies. sample_scalar
shows how to set it up.

SIMD on MSVC is not going to like it, though.